Optimize build time based on remote core type
Starting from release GLSDK 6.03.00.01 IPC needs to be
built for both M4 and C66x targets. M4 build is needed for
IPUMM while C66x builds are needed for DSP codecs.

If both the targets are enabled by default, IPC build takes a
long time (83 min for IPUMM) to complete the build. If the usecase is interested
in only one of the targets (M4 or C66x) this resulted in unnecessary
time needed to build the other target.

This patch optimizes this aspect by building only the target that is needed.

It brings in the following changes in build.

make ipumm: This will build IPC  only for M4 targets as needed for IPUMM.
make dspdce: Builds IPC only for C66x target as needed for DSP components.

In addition, if IPC needs to be built for specific targets the following commands
can be used:

make ipc_ipu
make ipc_dsp

In addition,
Instead of updating the products.mak file in IPC, environment variables are
supplied as command line arguments in make.

The build time after optimization is 12 min for IPUMM.

Signed-off-by: Pradeep Venkatasubbarao <pradeepv@ti.com>