uenv.txt: Use PARTUUID in SD boot args
Depending on the probe order of SD card, eMMC card and whether
eMMC is disabled on the board, the SD card gets registered as /dev/mmcblk0
or /dev/mmcblk1.
The behavior is different across boards and kernel.
Each time, we keep changing the root=/dev/mmcblkXp2 bootarg to match corerctly.

This can be avoided completely is using PARTUUID method.
PARTUUID is unique ID of the partition, which is different from the UUID assigned
by the filesystem.
Kernel can accept "root=PARTUUID=<partuuid value>" where the <partuuid> can
be populated by u-boot while generating the bootargs.

Change the bootargs to generate PARTUUID and populate into the bootargs

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>