Add lzop to the set of installed packages
This package is required to use lzo as the compresssion format for
kernel image instead of gzip.  lzo compression provides faster
decompression speed on the target compared to gzip while keeping the
size of the compressed binary roughly the same.

During my benchmarking, gzip yielded kernel of size 4.4 MB and
decompresssion time of 196 ms. lzo compression yielded kernel size of
4.6 MB and decompressions time of 69 ms.

Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>
Signed-off-by: Karthik Ramanan <a0393906@ti.com>